scripts/requirements.txt
beir
requests
psycopg2-binary